Workaround beast1 releases
Beast1 github started releasing "TempEST" in the  project. This means occasionally, 'latest' tag will be hogged by TempEST instead of beast. I have statically linked beast v1.10.4, latest as of this commit as a temporary workaround. Future replies to https://stackoverflow.com/questions/57687244/how-to-curl-github-api-release-and-filter-for-specific-string may see it return to a dynamic link.